Move layouts file
